Add guards to certain bits.h functions.
@craigbarnes recommended guarding against passing 0 to CLZ or CTZ, and
whilst perusing the functions I also noticed that round had a possible
divide-by-zero error waiting to happen, so I went ahead and added a
guard there as well. When the -DNDEBUG flag is not passed to a program
using these intrinsics, they should be prevented from crashing the
program and should be more easily diagnosable via the CE_ARG error state.